[NEW] spotify functions reorganization
