+++ src/helpers/listeners.js
@@ -1,12 +1,13 @@
+import { PERSONALITY } from "../commands/personality.js";
 import commands from "../commands/index.js";
+
 import {
   isAdmin,
   isCommand,
   reactionHandler,
   checkIsOnThread,
+  deleteSongFromPlaylist,
 } from "./index.js";
-
-import { PERSONALITY } from "../commands/personality.js";
 
 export const onPrivateMessage = async (message, client) => {
   const { author, content } = message;
@@ -93,5 +94,37 @@ export const removeReminder = (messageReaction, client, currentServer) => {
     } catch (err) {
       console.log("reminderError", err);
     }
+  }
+};
+
+export const removeSpotify = async (messageReaction, client, currentServer) => {
+  //remove song from client cache and spotify playlist using react
+  const { message, emoji, users } = messageReaction;
+  const { removeEmoji } = currentServer;
+
+  const foundMessageSpotify = client.playlistCachedMessages.find(
+    // found corresponding spotify message
+    ({ id }) => id === message.id
+  );
+
+  if (
+    process.env.USE_SPOTIFY === "yes" &&
+    foundMessageSpotify &&
+    emoji.name === removeEmoji &&
+    users.cache // if user reacting is the owner of spotify message
+      .map((user) => user.id)
+      .includes(message.mentions.users.first().id)
+  ) {
+    const { songId } = foundMessageSpotify;
+
+    const result = await deleteSongFromPlaylist(
+      songId,
+      client,
+      PERSONALITY.spotify
+    );
+    client.playlistCachedMessages = client.playlistCachedMessages.filter(
+      ({ id }) => id !== message.id
+    );
+    await message.reply(result);
   }
 };
